Addressing PR comments to simplify code use methods instead of abstractions
